app/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Exchanges & Blockchain
ccxt==4.2.25
web3==6.15.1

# Data & Analytics
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
pandas-ta==0.3.14b
ta==0.11.0
vectorbt==0.26.0

# ML & AI
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
torch==2.1.2
optuna==3.5.0
joblib==1.3.2
stable-baselines3==2.2.1
gymnasium==0.29.1

# Scheduling & Tasks
apscheduler==3.10.4
celery==5.3.4

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Utils
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
websockets==12.0
tenacity==8.2.3
aiohttp==3.9.1
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# Code Quality
black==23.12.1
ruff==0.1.11
mypy==1.8.0
pre-commit==3.6.0

# YAML
pyyaml==6.0.1

